Fixed task 0

+++ 0x1E-search_algorithms/0-linear.c
@@ -8,16 +8,16 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
+	int i;
 
 	if (array != NULL)
 	{
-		for (i = 0; i < size; i++)
+		for (i = 0; i < (int)size; i++)
 		{
-			printf("Value checked array[%zu] = [%d]\n", i, array[i]);
+			printf("Value checked array[%d] = [%d]\n", i, array[i]);
 			if (array[i] == value)
 			{
-				return ((int)i);
+				return (i);
 			}
 		}
 	}
